pyproject.toml
[project]
name = "to-naive-utc"
version = "0.1.0"
description = "Convert any datetime-like object in Python to naive UTC. No timezone - no problems"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Mark Lidenberg", email = "marklidenberg@gmail.com"},
]
license = {text = "MIT"}
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
    "build>=1.2.2.post1",
    "twine>=6.1.0",
    "pandas>=2.0.3",
    "numpy>=1.24.4",
    "pendulum>=3.0.0",
    "arrow>=1.3.0",
    "dony>=0.2.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
